+++ src/theme/colors.ts
@@ -34,6 +34,10 @@ export const colors = {
       background: 'rgba(255, 59, 48, 1)',
       text: whiteColor,
     },
+    scan: {
+      background: 'rgba(90, 200, 250, 1)',
+      text: whiteColor,
+    },
     switchEnv: {
       background: 'rgba(209, 209, 214, 1)',
     },