+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # v0.4.0
+* \#42 Dashboard Page knop toevoegen scannen
 
 # v0.3.0
 * \#54 Inloggen via LoginSite